ci: migrate to Poetry and modernise
Poetry is currently the most popular Python dependency manager. Whilst
it is ostenibly a replacement for the `requirements.txt` convention it is
also encourages more general best practices such as; dependency
lockfiles, releasing, task running, and so on.

This commit also introduces Ruff, a modern Python linter, that in our
case replaces our dependencies on `flake8` and `bandit`.